==> analysis/2_exclude.Rmd <==
---
title: "Action Asymmetry Pilot: Demographics & Exclusions"
pagetitle: "aah | pilot | exclusions"
author: "Owen Morgan"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document
---

<style type="text/css">
  body{
  font-family: Avenir;
  font-size: 12pt;
}
</style>

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE,
                      fig.align = "center", fig.width = 9,
                      fig.height = 6, results = "asis")
options(knitr.kable.NA = "")

cli.progress_show_after <- 0

# Suppress summarise info
options(dplyr.summarise.inform = FALSE)
```

```{r lib}
library(here)
library(tidyverse)
library(cli) # For printing error messages

library(gt)

source(here::here("lib", "load_process", "load_process.R"))
source(here::here("lib", "util.R"))
source(here::here("lib", "demographics.R"))
```

```{r config}
data_dir <- here::here("data")
proc_dir <- here::here(data_dir, "proc_pilot")
n_recruited = 112 ## Total recruited + paid. For reporting exclusions
```


##  {.tabset}
```{r load_combined, include = F}
## Load combined trial-per-row data
aah_task_all <- readr::read_tsv(
  here(proc_dir, "combined", "combined_task.tsv"),
  col_types = cols(
    subject = col_character(),
    time_elapsed_ms = col_double(),
    blocknum = col_double(),
    block_type = col_character(),
    block_response = col_character(),
    trialnum = col_double(),
    target = col_character(),
    level = col_character(),
    field = col_character(),
    target_present = col_character(),
    response = col_character(),
    correct = col_double(),
    rt = col_double(),
  )
) |>
  filter(block_type == "main" & target_present == "yes") |>
  select(subject,
         block_response,
         target,
         level,
         field,
         correct,
         rt)

aah_summary_all <- load_summary() |>
  select(
    subject,
    first_block,
    ehi_total,
    age,
    country,
    sex,
    education,
    race,
    hispanic_ethnicity,
    rt_overall,
    duration_s,
    task_experience_response,
    task_experience_other_response,
    open_ended_feedback_response,
    exclude_many_gos,
    exclude_low_acc,
    exclude_low_rt,
    exclude_high_rt
  ) |> rename(ehi = ehi_total)
```

### Demographics (All)
```{r demographics_summary}
rt_demo <- demo_summary_table(aah_summary_all)
rt_demo |> pretty_table() |> tab_header(title = "Demographics", subtitle = "Summary")
```
<br>
```{r country}
demo_country_table(aah_summary_all) |> pretty_table()
```
<br>
```{r race}
demo_race_table(aah_summary_all) |> pretty_table()
```
<br>
```{r ethnicity}
demo_ethnicity_table(aah_summary_all) |> pretty_table()
```
<br>
```{r task_experience}
demo_xp_table(aah_summary_all) |> pretty_table()
```
<br>

### Exclusions

```{r make_exclusions}
## Calculate any exclusions not based on task data.
##   (Exclusions based on task data are specified in:
##    1_process.Rmd, calling load_process.R/summarize_ind())
##       Responded "go" almost every time (78/80 or more)?
##       Accuracy at 60% or lower on any main block?
##       Median RT over 1500 or under 200?
##       Took longer than 45 minutes to do the task?
## Exclusions not based on task data:
##   Demographics exclusions that slipped through pre-screener:
##     Country (non-us); Age (not between 18 and 40)
##   Failed "Have you done this task before"
##   Has task data, but no EHI data. (e.g., if they quit inquisit after task)
##   (?) TODO: Failed interactive instructions

## Mark exclusion if country is not "US"
aah_summary_all <- aah_summary_all |>
  mutate(exclude_country = case_when(
    country == "US" ~ 0,
    TRUE ~ 1,)
    )

## Mark exclusion if age is not between 18 and 40
aah_summary_all <- aah_summary_all |>
  mutate(exclude_age = case_when(
    age >=18 & age <= 40 ~ 0,
    TRUE ~ 1
  ))

## Mark exclusion if failed "Have you done this task before?"
aah_summary_all <- aah_summary_all |> 
  mutate(exclude_done_before = case_when(
    task_experience_response == "No" ~ 0,
    task_experience_response == "Yes" ~ 1
  ))


## Mark exclusion if no EHI (if EHI is not between -100 & 100)
aah_summary_all <- aah_summary_all |>
  mutate(exclude_no_ehi = case_when(
    ehi >= -100 & ehi <= 100 ~ 0,
    TRUE ~ 1
  ))

## Update "exclude" column.
aah_summary_all <- aah_summary_all |>
  mutate(
    exclude = case_when(
      exclude_many_gos == 1 ~ 1,
      exclude_low_acc == 1 ~ 1,
      exclude_low_rt == 1 ~ 1,
      exclude_high_rt == 1 ~ 1,
      exclude_country == 1 ~ 1,
      exclude_age == 1 ~ 1,
      exclude_done_before == 1 ~ 1,
      exclude_no_ehi == 1 ~ 1,
      TRUE ~ 0
    )
  )

## Create long dataset with all subjects (included and excluded)
aah_all <- left_join(aah_task_all, aah_summary_all)

## Save long table with columns for all exclusion reasons.
## This is "the data."
readr::write_tsv(aah_all, here(proc_dir, "aah_long.tsv"))
readr::write_tsv(aah_summary_all, here(proc_dir, "aah_summary.tsv"))

## Create long data table with inclusions only
aah <- aah_all |> filter(exclude == 0)

## Create summary table with inclusions only
aah_summary <- aah_summary_all |> filter(exclude == 0)
```

```{r describe_exclusions_1}
n_all <- aah_summary_all$subject |> unique() |> length()
n_exclusions <- aah_summary_all$exclude |> sum()
n_keepers <- aah_summary_all |>
  filter(exclude == 0) |>
  (\(.) .[["subject"]])() |>
  unique() |>
  length()
tibble(
  "n (recruited)" = n_recruited,
  "n (with full task data)" = n_all,
  "exclusions" = n_exclusions,
  "n (keepers)" = n_keepers) |> 
  pretty_table()
```
<br>

We recruited `r n_recruited` right-handed participants through prolific. Of these, `r n_all` completed the task phase of the experiment. `r n_exclusions` were excluded for the reasons described below, leaving `r n_keepers` keepers.

<br>
```{r describe_exclusions_2}
exclusion_tbl <- aah_summary_all |> 
  rename(exclude_any = exclude) |> 
summarize(across(starts_with("exclude"), sum)) |> 
pivot_longer(cols = everything(), 
             names_to = 'exclusion', values_to = "n",
             names_prefix = "exclude_") |> 
  mutate('exclusion' = recode(exclusion,
    any = "Any",
    country = "Non-US country",
    age = "Age outside 18-40",
    done_before = "Reported having done the task before",
    high_rt = "High RT (median > 1500ms)",
    low_rt = "Low RT (median < 200ms)",
    low_acc = "Low accuracy (below 60% on any block)",
    many_gos = "Too many 'gos' (78/80 trials or more)",
    no_ehi = "Quit before completing EHI"
  )) |> 
  arrange(-n, exclusion) |> 
  select(n, exclusion) |> 
  rename('Reason for exclusion' = exclusion)

exclusion_tbl |> pretty_table()
```
<br>

### Demographics (Included)
```{r demographics_summary_included}
rt_demo <- demo_summary_table(aah_summary)
rt_demo |> pretty_table() |> tab_header(title = "Demographics", subtitle = "Summary")
```
<br>
```{r country_included}
demo_country_table(aah_summary) |> pretty_table()
```
<br>
```{r race_included}
demo_race_table(aah_summary) |> pretty_table()
```
<br>
```{r ethnicity_included}
demo_ethnicity_table(aah_summary) |> pretty_table()
```
<br>

### Feedback
```{r feedback}
feedback <- aah_summary |>
  filter(open_ended_feedback_response != "NA") |> 
  group_by(open_ended_feedback_response, exclude) |>
  rename(`Open-ended feedback` = open_ended_feedback_response,
         `Excluded?` = exclude) |> 
  summarize()
feedback |> pretty_table()
```
